Added DynamicValidationActionFilter (#1)
* Added `DynamicValidationActionFilter`.
* Added `DynamicValidationActionOperationFilter` to Swagger.

---

`DynamicValidationActionFilter` translates `BadRequestObjectResult` that has a value of type `SerializableError` to an `ErrorResponse` object, following the same convention as `DynamicValidationActionFilter` do. 

To make this work, you only need to return `BadRequest(ModelState);` in a controller action.

---

`DynamicValidationActionOperationFilter` add the `ErrorResponse` swagger definition to actions decorated by `[ProducesResponseType(400)]` that does not specify a type.